+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## main (unreleased)
 
+* [#99](https://github.com/Shopify/deprecation_toolkit/pull/99): Fix `Warning.warn` hook to accept a category.
 * [#95](https://github.com/Shopify/deprecation_toolkit/pull/95): Allow configuration of deprecation file paths and file names.
 
 ## 2.0.4 (2023-11-20)